Add checks to see if KFHit lands in sensor
This was used for the efficiency (TSOS contained). There were two options to define if a point is within a cell. The first one was to simply check if the closest detid to the point is actually the detid of the measurement used in the update step. The other method is to check using the inSensorCell() method which relies on the corners of the cells. These two methods do not produce the same results.